Module raytracing_n::utils

utils function and structures

Modules

number

module containing generic number.

Structs

NonNegative

A structure that guarenteen that the stored that is not negative, i.e >=0.

Normed

A structure that guarenteen that the stored that is normed.

Positive

A structure that guarantee that the stored that is positive, i.e >0.

Enums

ErrorSet

Error type return by set type functions.

Never

A type that can never be (safely) initialized. This is temporary, until never is accepted into stable rust.

Traits

Norm

A trait to get a norm a a element.