Trait raytracing_n::utils::Norm

pub trait Norm {
    type Output;
    fn norm(&self) -> Self::Output;
}

A trait to get a norm a a element.

A norm p(x) is defined by the following properties:

p(x + y) <= p(x) + p(y) for all x,y

p(sx) = |s|p(x) for all x and scalar s

p(x) = 0 then x = 0

Associated Types

type Output

The output type of the norm.

Required methods

fn norm(&self) -> Self::Output

Get the norm of the element.

The norm should return a non negative number, (i,e >= 0). In the future I may change to NonNegative.